refactor: 글 목록 및 댓글 조회 API 분리
