+++ api/kahlua/post.ts
@@ -1,6 +1,36 @@
 import { authInstance } from '@/api/auth/axios';
 import { MyPost } from '@/types/post';
 
+// 글 목록 조회
+export const fetchPostList = async ({
+  postType,
+  page,
+  size,
+  searchWord,
+}: {
+  postType: 'NOTICE' | 'KAHLUA_TIME';
+  page: number;
+  size: number;
+  searchWord?: string;
+}) => {
+  const response = await authInstance.get('/post/list', {
+    params: {
+      post_type: postType,
+      page,
+      size,
+      search_word: searchWord,
+    },
+  });
+
+  return response.data.result;
+};
+
+// 글 댓글 목록 조회
+export const fetchPostComments = async (postId: number) => {
+  const response = await authInstance.get(`/comment/${postId}/list`);
+  return response.data.result.comments || [];
+};
+
 // 내가 쓴 글 리스트 조회
 export const fetchMyPosts = async (
   page: number,
@@ -23,8 +53,7 @@ export const fetchMyPosts = async (
 // 댓글 수 조회
 export const fetchCommentCount = async (postId: number): Promise<number> => {
   try {
-    const response = await authInstance.get(`comment/${postId}/list`);
-    const comments = response.data.result.comments;
+    const comments = await fetchPostComments(postId);
     return comments.filter((comment: any) => comment.deletedAt === null).length;
   } catch (error) {
     console.error('댓글 수 조회 실패:', error);
