+++ components/mypage/ReservationList.tsx
@@ -3,7 +3,7 @@ import ButtonModal from '@/components/ui/ButtonModal';
 import {
   cancelReservation,
   getReservationList,
-} from '@/api/mypage/reservation';
+} from '@/api/kahlua/reservation';
 import { ReservationResponse } from '@/types/reservation';
 
 // 동방 예약 내역 리스트
